Crate generics

Structs

MyStructExample 🔒

Example Struct

TreasureChest 🔒

Struct with Generic

Enums

StatusMessage 🔒

Generic in enum

Functions

identity 🔒

identity<T> which has generic type Breakdown of the above code Generic is a place holder for the Type We specify as a generic type near function name We specify the same T to the type of the value Finally we returns the same T as return type This will accept any type of value to be used as value parameter while invoking them Note that T is common community convention- we can use any Name there for the generic

identity_f64 🔒

Identity with return value as i64

identity_i32 🔒

Identity with return value as i32

main 🔒

make_tuple 🔒

function may have multiple arguments

make_tuple_both_generic 🔒

Multiple Generics - function may have multiple generic

make_tuple_with_generic 🔒

Multiple argument Generics - function may have multiple arg with one generic