generics

Function identity

pub(crate) fn identity<T>(value: T) -> T

identity<T> which has generic type Breakdown of the above code Generic is a place holder for the Type We specify as a generic type near function name We specify the same T to the type of the value Finally we returns the same T as return type This will accept any type of value to be used as value parameter while invoking them Note that T is common community convention- we can use any Name there for the generic